REFACTORED: Unit tests of adapted colors
This commit provides the set of unit tests that was refactored to cover colors adapted so that there is no need in many simulators any more.

Now it requires iOS 13 or higher for running unit tests.

REFACTORED: The set of unit test to cover [SemanticColorsAdapted] and [SystemColorsAdapted] to reduce required simulators for testing to only one.